moving types files to _types
