+++ nextjs/src/app/_types/ticker-columns.tsx
