+++ nextjs/src/app/_types/ticker.ts
